space/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pii-masking-space"
version = "0.1.0"
description = "PII Masking Demo Space"
requires-python = ">=3.9"
dependencies = [
    # FastAPI and web server
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    
    # HTTP Client for API calls
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
    
    # Mistral AI client
    "mistralai>=1.0.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    
    # Basic data processing
    "numpy>=1.24.0",
    
    # BERT and ML dependencies
    "torch>=2.0.0,<2.3.0",
    "transformers>=4.35.0",
    "tokenizers>=0.15.0",
]

[tool.uv]
dev-dependencies = []
